HeatMap component is split into multiple components
